combined feather and weather, now weather fetches properly
